change hand-util.js to hand-util.ts
