+++ src/main/ts/jan/knitted-straight.ts
@@ -74,4 +74,5 @@ export const KNITTED_STRAIGHT = {
 } as const;
 
 export type KnittedStraight = typeof KNITTED_STRAIGHT[keyof typeof KNITTED_STRAIGHT];
-export type KnittedType = (keyof typeof KNITTED_STRAIGHT)[];
+export type KnittedType = keyof typeof KNITTED_STRAIGHT;
+export type KnittedTypes = (keyof typeof KNITTED_STRAIGHT)[];
