+++ src/main/ts/jan/mcr-util.js
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
  */
-import HandUtil from './hand-util';
+import * as HandUtil from './hand-util';
 import {hasJanpai} from './jan-util';
 import {McrComplete} from './mcr-complete';
 import {MCR_YAKU} from './mcr-yaku';
